Move test_sphere_coords to new file, test_scatteringtheory
