+++ holopy/scattering/tests/test_basics.py
@@ -24,21 +24,12 @@ from ..scatterer import Sphere, Difference
 from ...core import detector_grid, detector_points
 from ..theory import Mie
 from ...core.tests.common import assert_obj_close
-from ...core.metadata import sphere_coords, update_metadata
+from ...core.metadata import update_metadata
 from ..calculations import calc_intensity, calc_holo, calc_field
 from ..theory.scatteringtheory import stack_spherical
 
 # small tests against results from the previous version of holopy
 
-@attr("fast")
-def test_sphere_coords():
-    t = detector_grid(shape = (2,2), spacing = .1)
-    p = sphere_coords(t, wavevec=2*np.pi*1.33/.66, origin=(0,0,1))
-    pos = stack_spherical(p).T
-    assert_allclose(pos, np.array([[ 12.66157039,   0.        ,   0.        ],
-       [ 12.72472076,   0.09966865,   1.57079633],
-       [ 12.72472076,   0.09966865,   0.        ],
-       [ 12.78755927,   0.1404897 ,   0.78539816]]))
 
 @attr("fast")
 def test_calc_field():
